@@ -1,9 +1,8 @@
-[tool.black]
-line-length = 80
-skip-string-normalization = true
-
 [tool.ruff]
-target-version = "py37"
+line-length = 80
+target-version = "py38"
+
+[tool.ruff.lint]
 select = [
     # Pycodestyle
     "E",
@@ -18,12 +17,16 @@ select = [
     # Comprehensions
     "C4",
 
+    # Debugger
+    "T100",
+
     # Pyupgrade
     "UP",
 
     # NumPy-specific rules
     "NPY001",
     "NPY003",
+    "NPY201",
 
     # Bugbear
     "B002",
@@ -60,6 +63,9 @@ select = [
 
     # Pygrep-hooks
     "PGH004",
+
+    # Ruff-specific rules
+    "RUF100",
 ]
 unfixable = [
     "NPY001"
@@ -74,3 +80,7 @@ ignore = [
     "UP015",
     "UP038"
 ]
+
+[tool.ruff.lint.isort]
+combine-as-imports = true
+known-first-party = ["paddle"]
