@@ -15,47 +15,47 @@ breathe_default_project = "PaddlePaddle"
 MarkdownParser = parser.CommonMarkParser
 AutoStructify = transform.AutoStructify
 AutoStructify.default_config = {
-    'enable_auto_doc_ref': False,
-    'auto_toc_maxdepth': 1,
-    'auto_toc_tree_section': None,
-    'enable_auto_toc_tree': True,
-    'enable_eval_rst': True,
-    'enable_math': True,
-    'enable_inline_math': True,
-    'commonmark_suffixes': ['.md'],
-    'url_resolver': lambda x: x,
-    'known_url_schemes': ['http', 'https'],
+    "enable_auto_doc_ref": False,
+    "auto_toc_maxdepth": 1,
+    "auto_toc_tree_section": None,
+    "enable_auto_toc_tree": True,
+    "enable_eval_rst": True,
+    "enable_math": True,
+    "enable_inline_math": True,
+    "commonmark_suffixes": [".md"],
+    "url_resolver": lambda x: x,
+    "known_url_schemes": ["http", "https"],
 }
 
 templates_path = ["/templates"]
 
-project = 'PaddlePaddle'
-author = '%s developers' % project
-copyright = '%d, %s' % (time.localtime(time.time()).tm_year, author)
-github_doc_root = 'https://github.com/PaddlePaddle/docs/docs'
+project = "PaddlePaddle"
+author = "%s developers" % project
+copyright = "%d, %s" % (time.localtime(time.time()).tm_year, author)
+github_doc_root = "https://github.com/PaddlePaddle/docs/docs"
 
 # add markdown parser
 MarkdownParser.github_doc_root = github_doc_root
 
-os.environ['PADDLE_BUILD_DOC'] = '1'
+os.environ["PADDLE_BUILD_DOC"] = "1"
 
 # Add any Sphinx extension moduleexclude_patterns names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom ones
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.autosummary',
-    'sphinx.ext.mathjax',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.graphviz',
-    'sphinx_sitemap',
-    'sphinx.ext.linkcode',
-    'recommonmark',
-    'sphinx_markdown_tables',
-    'breathe',
-    'exhale',
-    'sphinx.ext.autosectionlabel',
-    'nbsphinx',
-    'sphinx_design',
+    "sphinx.ext.autodoc",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.mathjax",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.graphviz",
+    "sphinx_sitemap",
+    "sphinx.ext.linkcode",
+    "recommonmark",
+    "sphinx_markdown_tables",
+    "breathe",
+    "exhale",
+    "sphinx.ext.autosectionlabel",
+    "nbsphinx",
+    "sphinx_design",
 ]
 
 # exhale
@@ -90,47 +90,47 @@ cpp_paren_attributes = ["optimize", "__aligned__", "section", "deprecated"]
 # primary_domain="cpp"
 
 # nbsphinx
-nbsphinx_execute = 'never'
+nbsphinx_execute = "never"
 
 # math
 math_number_all = False
 mathjax_config = {
-    'extensions': ['tex2jax.js'],
-    'jax': ['input/TeX', 'output/HTML-CSS'],
-    'tex2jax': {
-        'inlineMath': [['$', '$'], ["\\(", "\\)"]],
-        'processEscapes': 'false',
+    "extensions": ["tex2jax.js"],
+    "jax": ["input/TeX", "output/HTML-CSS"],
+    "tex2jax": {
+        "inlineMath": [["$", "$"], ["\\(", "\\)"]],
+        "processEscapes": "false",
     },
 }
 mathjax_path = "https://cdn.bootcss.com/mathjax/2.7.6/MathJax.js"
 
 table_styling_embed_css = True
 
-autodoc_member_order = 'bysource'
+autodoc_member_order = "bysource"
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 # source_suffix = ['.rst', '.md']
 source_suffix = {
-    '.rst': 'restructuredtext',
+    ".rst": "restructuredtext",
     #'.txt': 'markdown',
-    '.md': 'markdown',
+    ".md": "markdown",
 }
 
 # The encoding of source files.
-source_encoding = 'utf-8'
+source_encoding = "utf-8"
 
 # The master toctree document.
-master_doc = 'index_cn'
+master_doc = "index_cn"
 
-html_baseurl = 'https://www.paddlepaddle.org.cn/documentation/docs/'
+html_baseurl = "https://www.paddlepaddle.org.cn/documentation/docs/"
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = 'zh'
-version = ''
+language = "zh"
+version = ""
 # There are two options for replacing |today|: either, you set today to some
 # non-false value, then it is used:
 # today = ''
@@ -142,14 +142,14 @@ version = ''
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 exclude_patterns = [
-    '_build',
-    'book/*',
-    'design/*',
-    '**/*_en.rst',
-    '*_en.rst',
-    '**/*hidden.*',
-    '**/*.en*',
-    '*.en*',
+    "_build",
+    "book/*",
+    "design/*",
+    "**/*_en.rst",
+    "*_en.rst",
+    "**/*hidden.*",
+    "**/*.en*",
+    "*.en*",
     "**/*CMakeLists**",
 ]
 
@@ -169,7 +169,7 @@ exclude_patterns = [
 # show_authors = False
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
+pygments_style = "sphinx"
 
 # A list of ignored prefixes for module index sorting.
 # modindex_common_prefix = []
@@ -180,59 +180,59 @@ pygments_style = 'sphinx'
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = False
 html_sidebars = {
-    '**': ['globaltoc.html'],
+    "**": ["globaltoc.html"],
 }
 
-doc_version = os.environ.get('VERSIONSTR', 'develop')
+doc_version = os.environ.get("VERSIONSTR", "develop")
 if len(doc_version) == 0:
-    doc_version = 'develop'
+    doc_version = "develop"
 
-GITHUB_REPO_URL = 'https://github.com/PaddlePaddle/Paddle/blob/'
-if doc_version != 'develop':
-    GITHUB_REPO_URL += 'release/'
+GITHUB_REPO_URL = "https://github.com/PaddlePaddle/Paddle/blob/"
+if doc_version != "develop":
+    GITHUB_REPO_URL += "release/"
 
 # -- Options for HTML output ----------------------------------------------
 html_context = {
-    'display_github': True,
-    'github_user': 'PaddlePaddle',
-    'github_repo': 'docs',
-    'github_version': 'develop',
-    'conf_py_path': '/docs/',
+    "display_github": True,
+    "github_user": "PaddlePaddle",
+    "github_repo": "docs",
+    "github_version": "develop",
+    "conf_py_path": "/docs/",
 }
 
-if 'VERSIONSTR' in os.environ and os.environ['VERSIONSTR'] != 'develop':
+if "VERSIONSTR" in os.environ and os.environ["VERSIONSTR"] != "develop":
     try:
-        float(os.environ['VERSIONSTR'])
-        html_context['github_version'] = 'release/' + os.environ['VERSIONSTR']
+        float(os.environ["VERSIONSTR"])
+        html_context["github_version"] = "release/" + os.environ["VERSIONSTR"]
     except ValueError:
         print(
             "os.environ['VERSIONSTR']={} is not releases's name".format(
-                os.environ['VERSIONSTR']
+                os.environ["VERSIONSTR"]
             )
         )
-        html_context['github_version'] = os.environ['VERSIONSTR']
+        html_context["github_version"] = os.environ["VERSIONSTR"]
 
 # if lang == 'en' and 'pagename' in html_context and html_context['pagename'].startswith('api/'):
 #     html_context['display_github'] = False
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
 html_theme_options = {
     # 'canonical_url': '',
     # 'analytics_id': 'UA-XXXXXXX-1',  #  Provided by Google in your dashboard
-    'logo_only': False,
-    'display_version': True,
-    'prev_next_buttons_location': 'bottom',
-    'style_external_links': True,
-    'vcs_pageview_mode': 'blob',
-    'style_nav_header_background': 'white',
+    "logo_only": False,
+    "display_version": True,
+    "prev_next_buttons_location": "bottom",
+    "style_external_links": True,
+    "vcs_pageview_mode": "blob",
+    "style_nav_header_background": "white",
     # Toc options
-    'collapse_navigation': False,
-    'sticky_navigation': True,
-    'navigation_depth': 10,
-    'includehidden': True,
-    'titles_only': True,
+    "collapse_navigation": False,
+    "sticky_navigation": True,
+    "navigation_depth": 10,
+    "includehidden": True,
+    "titles_only": True,
 }
 
 # Add any paths that contain custom static files (such as style sheets) here,
@@ -241,10 +241,10 @@ html_theme_options = {
 # html_static_path = []
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = project + 'doc'
+htmlhelp_basename = project + "doc"
 
 # -- Options for LaTeX output ---------------------------------------------
-latex_engine = 'xelatex'
+latex_engine = "xelatex"
 latex_elements = {
     # The paper size ('letterpaper' or 'a4paper').
     #
@@ -252,11 +252,11 @@ latex_elements = {
     # The font size ('10pt', '11pt' or '12pt').
     #
     # 'pointsize': '10pt',
-    'fncychap': '',
+    "fncychap": "",
     # Additional stuff for the LaTeX preamble.
     #
-    'preamble': r'''\usepackage{ctex}
-    ''',
+    "preamble": r"""\usepackage{ctex}
+    """,
     # Latex figure (float) alignment
     #
     # 'figure_align': 'htbp',
@@ -266,20 +266,20 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, '%s.tex' % project, project, author, 'manual'),
+    (master_doc, "%s.tex" % project, project, author, "manual"),
 ]
 
 # Use the .. admonition:: directive for Notes sections.
 # False to use the .. rubric:: directive instead.
 napoleon_use_admonition_for_notes = True
 numfig = False
-highlight_language = 'python'
+highlight_language = "python"
 html_experimental_html5_writer = False
 MARKDOWN_EXTENSIONS = [
-    'markdown.extensions.fenced_code',
-    'markdown.extensions.tables',
-    'pymdownx.superfences',
-    'pymdownx.escapeall',
+    "markdown.extensions.fenced_code",
+    "markdown.extensions.tables",
+    "pymdownx.superfences",
+    "pymdownx.escapeall",
 ]
 
 math_numfig = False
@@ -306,34 +306,34 @@ def linkcode_resolve(domain, info):
     :param info:
     :return:
     """
-    if domain != 'py':
+    if domain != "py":
         return None
-    if not info['fullname']:
+    if not info["fullname"]:
         return None
-    filename = info['fullname'].replace('.', '/')
-    class_names = info['fullname'].split('.')
+    filename = info["fullname"].replace(".", "/")
+    class_names = info["fullname"].split(".")
     api_title = class_names[len(class_names) - 1]
     print(api_title)
-    class_name = info['fullname'].replace('.' + api_title, '')
+    class_name = info["fullname"].replace("." + api_title, "")
     try:
         current_class = sys.modules[class_name]
         api = getattr(current_class, api_title)
         line_no = None
 
-        if type(api).__name__ == 'module':
-            module = os.path.splitext(api.__file__)[0] + '.py'
+        if type(api).__name__ == "module":
+            module = os.path.splitext(api.__file__)[0] + ".py"
         else:
             node_definition = (
                 ast.ClassDef if inspect.isclass(api) else ast.FunctionDef
             )
 
             if api.__module__ not in [
-                'paddle.fluid.core',
-                'paddle.fluid.layers.layer_function_generator',
+                "paddle.fluid.core",
+                "paddle.fluid.layers.layer_function_generator",
             ]:
                 module = (
                     os.path.splitext(sys.modules[api.__module__].__file__)[0]
-                    + '.py'
+                    + ".py"
                 )
                 with open(module) as module_file:
                     module_ast = ast.parse(module_file.read())
@@ -355,12 +355,12 @@ def linkcode_resolve(domain, info):
                                 line_no = node.lineno
                                 break
             else:
-                module = os.path.splitext(current_class.__file__)[0] + '.py'
+                module = os.path.splitext(current_class.__file__)[0] + ".py"
         url = GITHUB_REPO_URL + os.path.join(
-            doc_version, 'python', module[module.rfind('paddle') :]
+            doc_version, "python", module[module.rfind("paddle") :]
         )
         if line_no:
-            return url + '#L' + str(line_no)
+            return url + "#L" + str(line_no)
         return url
     except Exception as e:
         print("conf.py linkcode_resolve error", e)
@@ -376,15 +376,15 @@ def setup(app):
     # Add hook for building doxygen xml when needed
     # no c++ API for now
     app.add_config_value(
-        'recommonmark_config',
+        "recommonmark_config",
         {
             # 'url_resolver': lambda url: github_doc_root + url,
-            'enable_math': True,
-            'enable_inline_math': True,
-            'enable_eval_rst': True,
-            'enable_auto_doc_ref': True,
-            'auto_toc_tree_section': True,
-            'known_url_schemes': ['http', 'https'],
+            "enable_math": True,
+            "enable_inline_math": True,
+            "enable_eval_rst": True,
+            "enable_auto_doc_ref": True,
+            "auto_toc_tree_section": True,
+            "known_url_schemes": ["http", "https"],
         },
         True,
     )
