@@ -93,7 +93,7 @@ def extract_code_blocks_from_rst(docstr):
                     if cb_cur_indent <= mo.start():
                         cb_cur.append(linecont)
                     else:
-                        if linecont[mo.start()] == '#':
+                        if linecont[mo.start()] == "#":
                             continue
                         else:
                             # block end
@@ -104,7 +104,7 @@ def extract_code_blocks_from_rst(docstr):
                             cb_started = False
                             cb_cur_indent = -1
                             cb_cur = []
-    logger.info('extracted %d code blocks.', len(code_blocks))
+    logger.info("extracted %d code blocks.", len(code_blocks))
     return code_blocks
 
 
@@ -123,10 +123,10 @@ def extract_code_blocks_from_md(docstr):
 def extract_code_blocks_from_file(filename):
     r = os.path.splitext(filename)
     ext = r[1].lower()
-    if ext == '.md':
-        return extract_code_blocks_from_md(open(filename, 'r').read())
-    elif ext == '.rst':
-        return extract_code_blocks_from_rst(open(filename, 'r').read())
+    if ext == ".md":
+        return extract_code_blocks_from_md(open(filename, "r").read())
+    elif ext == ".rst":
+        return extract_code_blocks_from_rst(open(filename, "r").read())
     else:
         return []
 
@@ -139,10 +139,10 @@ def get_all_files(p):
     for path, dirs, files in os.walk(p):
         for filename in files:
             r = os.path.splitext(filename)
-            logger.info('%s found', filename)
-            if len(r) == 2 and r[1].lower() in ['.md', '.rst']:
+            logger.info("%s found", filename)
+            if len(r) == 2 and r[1].lower() in [".md", ".rst"]:
                 filelist.append(os.path.join(path, filename))
-    logger.info('find %d files from %s.', len(filelist), p)
+    logger.info("find %d files from %s.", len(filelist), p)
     return filelist
 
 
@@ -158,9 +158,9 @@ def find_all_paddle_api_from_code_block(cbstr):
     # case 4: in docstring
     api_set = set()
     ds_list = cbstr.split("\n")
-    import_pat = re.compile(r'from\s+([\.\w]+)\s+import\s+(\w+)')
-    normal_pat = re.compile(r'(paddle\.[\.\w]+)')
-    docstr_pat = re.compile(r'((\'{3})|(\"{3}))')
+    import_pat = re.compile(r"from\s+([\.\w]+)\s+import\s+(\w+)")
+    normal_pat = re.compile(r"(paddle\.[\.\w]+)")
+    docstr_pat = re.compile(r"((\'{3})|(\"{3}))")
     in_docstr = False
     for line in ds_list:
         line = line.strip()
@@ -168,13 +168,13 @@ def find_all_paddle_api_from_code_block(cbstr):
             in_docstr = not in_docstr
         if in_docstr:
             continue
-        sharp_ind = line.find('#')
+        sharp_ind = line.find("#")
         mo_i = import_pat.search(line)
         if mo_i:
             if (sharp_ind < 0 or mo_i.start() < sharp_ind) and mo_i.group(
                 1
-            ).startswith('paddle'):
-                api_set.add(f'{mo_i.group(1)}.{mo_i.group(2)}')
+            ).startswith("paddle"):
+                api_set.add(f"{mo_i.group(1)}.{mo_i.group(2)}")
         else:
             mo_n = normal_pat.finditer(line)
             for mo in mo_n:
@@ -186,10 +186,10 @@ def find_all_paddle_api_from_code_block(cbstr):
 def extract_api_from_file(filename):
     api_set = set()
     codeblocks = extract_code_blocks_from_file(filename)
-    logger.info('find %d code-blocks from %s.', len(codeblocks), filename)
+    logger.info("find %d code-blocks from %s.", len(codeblocks), filename)
     for cb in codeblocks:
         api_set.update(find_all_paddle_api_from_code_block(cb))
-    logger.info('find %d apis from %s.', len(api_set), filename)
+    logger.info("find %d apis from %s.", len(api_set), filename)
     return api_set
 
 
@@ -197,9 +197,9 @@ def extract_doc_title_from_file(filename):
     r = os.path.splitext(filename)
     if len(r) != 2:
         return None
-    if r[1].lower() == '.md':
+    if r[1].lower() == ".md":
         return extract_md_title(filename)
-    elif r[1].lower() == '.rst':
+    elif r[1].lower() == ".rst":
         return extract_rst_title(filename)
     return None
 
@@ -220,10 +220,10 @@ def extract_rst_title(filename):
     overrides = {
         # Disable the promotion of a lone top-level section title to document
         # title (and subsequent section title to document subtitle promotion).
-        'docinfo_xform': 0,
-        'initial_header_level': 2,
+        "docinfo_xform": 0,
+        "initial_header_level": 2,
     }
-    with open(filename, 'r') as fileobj:
+    with open(filename, "r") as fileobj:
         doctree = docutils.core.publish_doctree(
             fileobj.read(), settings_overrides=overrides
         )
@@ -235,14 +235,14 @@ def extract_rst_title(filename):
     return None
 
 
-def extract_params_desc_from_rst_file(filename, section_title='参数'):
+def extract_params_desc_from_rst_file(filename, section_title="参数"):
     overrides = {
         # Disable the promotion of a lone top-level section title to document
         # title (and subsequent section title to document subtitle promotion).
-        'docinfo_xform': 0,
-        'initial_header_level': 2,
+        "docinfo_xform": 0,
+        "initial_header_level": 2,
     }
-    with open(filename, 'r') as fileobj:
+    with open(filename, "r") as fileobj:
         doctree = docutils.core.publish_doctree(
             fileobj.read(), settings_overrides=overrides
         )
@@ -266,11 +266,11 @@ def extract_params_desc_from_rst_file(filename, section_title='参数'):
 
 
 def extract_md_title(filename):
-    with open(filename, 'r') as fileobj:
+    with open(filename, "r") as fileobj:
         html = markdown.markdown(fileobj.read())
-        mo = re.search(r'<h1>(.*?)</h1>', html)
+        mo = re.search(r"<h1>(.*?)</h1>", html)
         if mo:
-            mos = re.search(r'<strong>(.*?)</strong>', mo.group(1))
+            mos = re.search(r"<strong>(.*?)</strong>", mo.group(1))
             return mos.group(1) if mos else mo.group(1)
     return None
 
@@ -283,7 +283,7 @@ def format_filename(filename):
     function get_all_files does not format it.
     """
     rp = os.path.realpath(filename)
-    pat_str = 'docs/'  # if the structure changed, update this pattern
+    pat_str = "docs/"  # if the structure changed, update this pattern
     ind = rp.rindex(pat_str)
     if ind >= 0:
         return rp[ind + len(pat_str) :]
@@ -305,21 +305,21 @@ def extract_all_infos(docdirs):
 
 
 def ref_role(role, rawtext, text, lineno, inliner, options=None, content=None):
-    '''dummy ref role'''
+    """dummy ref role"""
     ref_target = text
     node = docutils.nodes.reference(rawtext, text)
     return [node], []
 
 
-docutils.parsers.rst.roles.register_canonical_role('ref', ref_role)
-docutils.parsers.rst.roles.register_local_role('ref', ref_role)
+docutils.parsers.rst.roles.register_canonical_role("ref", ref_role)
+docutils.parsers.rst.roles.register_local_role("ref", ref_role)
 
 
 class PyFunctionDirective(docutils.parsers.rst.Directive):
-    '''dummy py:function directive
+    """dummy py:function directive
 
     see https://docutils-zh-cn.readthedocs.io/zh_CN/latest/howto/rst-roles.html
-    '''
+    """
 
     required_arguments = 0
     optional_arguments = 0
@@ -328,24 +328,24 @@ class PyFunctionDirective(docutils.parsers.rst.Directive):
     has_content = True
 
     def run(self):
-        text = '\n'.join(self.content)
+        text = "\n".join(self.content)
         thenode = docutils.nodes.title(text, text)
         return [thenode]
 
 
 docutils.parsers.rst.directives.register_directive(
-    'py:function', PyFunctionDirective
+    "py:function", PyFunctionDirective
 )  # as abs_cn.rst
 docutils.parsers.rst.directives.register_directive(
-    'py:class', PyFunctionDirective
+    "py:class", PyFunctionDirective
 )  # as Tensor_cn.rst
 docutils.parsers.rst.directives.register_directive(
-    'py:method', PyFunctionDirective
+    "py:method", PyFunctionDirective
 )  # as grad_cn.rst
 
 
 class ToctreeDirective(docutils.parsers.rst.Directive):
-    '''dummy toctree directive'''
+    """dummy toctree directive"""
 
     required_arguments = 1
     optional_arguments = 5
@@ -357,16 +357,16 @@ class ToctreeDirective(docutils.parsers.rst.Directive):
         return []
 
 
-docutils.parsers.rst.directives.register_directive('toctree', ToctreeDirective)
+docutils.parsers.rst.directives.register_directive("toctree", ToctreeDirective)
 
 arguments = [
     # flags, dest, type, default, help
     [
-        '--output',
-        'output',
+        "--output",
+        "output",
         str,
-        'called_apis_from_docs.json',
-        'output filename. default: called_apis_from_docs.json',
+        "called_apis_from_docs.json",
+        "output filename. default: called_apis_from_docs.json",
     ],
 ]
 
@@ -377,14 +377,14 @@ def parse_args():
     """
     global arguments
     parser = argparse.ArgumentParser(
-        description='extract all the called apis from md or reST files.'
+        description="extract all the called apis from md or reST files."
     )
     parser.add_argument(
-        'dir',
+        "dir",
         type=str,
-        help='travel all the files include this directory',
-        default='.',
-        nargs='+',
+        help="travel all the files include this directory",
+        default=".",
+        nargs="+",
     )
     for item in arguments:
         parser.add_argument(
@@ -402,10 +402,10 @@ if __name__ == "__main__":
     apis_dict, file_titles = extract_all_infos(args.dir)
     import json
 
-    with open(args.output, 'w') as f:
+    with open(args.output, "w") as f:
         json.dump(apis_dict, f, indent=4)
     r = os.path.splitext(args.output)
-    with open(f'{r[0]}-titles{r[1]}', 'w') as f:
+    with open(f"{r[0]}-titles{r[1]}", "w") as f:
         json.dump(file_titles, f, indent=4)
 
-    print('Done')
+    print("Done")
